Move ObjectStoragePath and attach to Task SDK
These things don't actually work without Core, which should be available
at runtime. But you can import the class and function and call them at
the top-level.
