+++ providers/common/io/tests/system/common/io/example_file_transfer_local_to_s3.py
@@ -23,9 +23,14 @@ from typing import cast
 
 from airflow import DAG
 from airflow.decorators import task
-from airflow.io.path import ObjectStoragePath
 from airflow.providers.common.io.operators.file_transfer import FileTransferOperator
+from airflow.providers.common.io.version_compat import AIRFLOW_V_3_0_PLUS
 from airflow.utils.trigger_rule import TriggerRule
+
+if AIRFLOW_V_3_0_PLUS:
+    from airflow.sdk import ObjectStoragePath
+else:
+    from airflow.io.path import ObjectStoragePath  # type: ignore[no-redef]
 
 ENV_ID = os.environ.get("SYSTEM_TESTS_ENV_ID")
 DAG_ID = "example_file_transfer_local_to_s3"
