+++ task-sdk/pyproject.toml
@@ -24,6 +24,7 @@ requires-python = ">=3.9, <3.13"
 dependencies = [
     "aiologic>=0.14.0",
     "attrs>=24.2.0, !=25.2.0",
+    "fsspec>=2023.10.0",
     "httpx>=0.27.0",
     "jinja2>=3.1.5",
     "methodtools>=0.4.7",
